+++ lib/greeklish_iso843.rb
@@ -1,3 +1,3 @@
 module GreeklishIso843; end
 
-require 'greeklish_iso843/converter'
+require 'greeklish_iso843/greek_text'