+++ lib/greeklish_iso843/core_ext/string.rb
@@ -2,6 +2,6 @@ require 'greeklish_iso843'
 
 class String
   def to_greeklish
-    GreeklishIso843::Converter.convert(self)
+    GreeklishIso843::GreekText.to_greeklish(self)
   end
 end